+++ src/store/reducers/authentication/types.ts
@@ -1,10 +1,3 @@
 export interface AuthenticationState {
     isAuthenticated: boolean;
-    isUserAllowedToMakeChanges?: boolean;
-    isViewerUser?: boolean;
-
-    user: string | undefined;
-    id: string | undefined;
-
-    metaUser: string | undefined;
 }